Day 25, create Primes!
In this challenge, you will determine if a given number X is prime or not. A prime number is a natural number greater than 1 that has no positive divisors other than 1 and itself. You will be given N numbers and for each, you will print out "Prime" if the number is prime or "Not prime" if the number is not prime.

Input Format
The first line of the input is T, total number of test cases. Each of the next line contains an integer N.

Constraints
1≤T≤20
1≤N≤2×109

Output Format
For each test case print Prime if N is prime, otherwise print Not prime.

Sample Input
3
12
5
7

Sample Output
Not prime
Prime
Prime